agent_engine/requirements.txt
fastapi
uvicorn[standard]
crewai
langchain-google-genai
google-generativeai
playwright
pandas
aiohttp
python-dotenv
websockets
pillow
httpx
pydantic
numpy
scikit-learn
redis
rq
python-multipart
slowapi
moviepy
imageio-ffmpeg
pysubs2
google-genai
